+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 When upgrading to a new version, make sure to follow the directions under the "Upgrading" header of the corresponding version.
 If there is no "Upgrading" header for that version, no post-upgrade actions need to be performed.
 
+## Upcoming
+
+### New Features
+- This project is now licensed under the GPLv3 license instead of the MIT
+license
+
+
 ## 4.1 (2023-06-04)
 
 ### Improvements